test: split compare complex kernel.
